Conflicts:
	CHANGELOG.md
	features/Context/FixturesContext.php
	spec/Pim/Bundle/CatalogBundle/Builder/ProductBuilderSpec.php
	src/Pim/Bundle/CatalogBundle/Entity/Group.php
	src/Pim/Bundle/CatalogBundle/Manager/MediaManager.php

+++ src/Pim/Bundle/EnrichBundle/Resources/public/js/pim-attributeoptionview.js
@@ -138,7 +138,9 @@ define(
                 }
             },
             deleteItem: function() {
-                this.parent.deleteItem(this);
+                Dialog.confirm(__('pim_enrich.item.delete.confirm.content'), __('pim_enrich.item.delete.confirm.title'), _.bind(function () {
+                    this.parent.deleteItem(this);
+                }, this));
             },
             updateItem: function() {
                 this.inLoading(true);
